changed BookingController and UserController
BookingController : add user picture in showBookingById
UserController : changed updatePicture

+++ app/Http/Controllers/UserController.php
@@ -283,8 +283,9 @@ class UserController extends Controller
         if($validator->fails()){
             return response()->json($validator->errors()->toJson(), 400);
         }
-    
-        unlink('storage/'.$user->user_picture);
+        if($user->user_picture != null){
+            unlink('storage/'.$user->user_picture);
+        }
         $file = $request->file('user_picture');
         $upload_dest = 'user_picture';
         $extension = $file->extension();
